Refactor test for CSE
